feat(Util): Util 생성
로또에만 제한되게 사용할 수 있는 것이 아닌 util들을 모아 놓은 Util 클래스를 작성했습니다.
자연수 배수인지 체크해주는 validateMultiplesOf, 몇 배수인지 계산해주는 getHowManyTimes 메소드들을 Util로 이동시켰습니다.
