+++ src/main/java/lotto/utils/LottoUtil.java
@@ -1,6 +1,5 @@
 package lotto.utils;
 
-import static lotto.message.ErrorMessage.AMOUNT_IS_NOT_IN_THOUSAND_WON_UNITS;
 import static lotto.message.ErrorMessage.LOTTO_NUMBER_IS_NOT_IN_RANGE;
 
 import camp.nextstep.edu.missionutils.Randoms;
@@ -30,16 +29,5 @@ public class LottoUtil {
 
     public static Lotto generateLotto() {
         return new Lotto(Randoms.pickUniqueNumbersInRange(1, 45, 6));
-    }
-
-    public static BigDecimal getNumberOfLottoPurchased(BigDecimal purchaseAmount, BigDecimal lottoPrice) {
-        validateMultiplesOf(purchaseAmount, lottoPrice);
-        return purchaseAmount.divide(lottoPrice, BigDecimal.ROUND_DOWN);
-    }
-
-    public static void validateMultiplesOf(BigDecimal num, BigDecimal factor) {
-        if (num.remainder(factor).compareTo(BigDecimal.ZERO) != 0) {
-            throw new IllegalArgumentException(AMOUNT_IS_NOT_IN_THOUSAND_WON_UNITS.getMessage());
-        }
     }
 }
